+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## 5.6.0 - 2026-04-24
 
 ### Added
-- Added a health check that calls the `backup/health` craft command to monitor backups managed by the [Craft Backup](https://github.com/webhubworks/craft-backup) plugin.
+- Added a health check that calls the `backup/monitor` craft command to monitor backups managed by the [Craft Backup](https://github.com/webhubworks/craft-backup) plugin.
 
 ## 5.5.0 - 2026-04-20